modules: Make module uncompression explicit
Up to now we have implicitly uncompressed every
module. This is counterproductive for cases where
the module shall be compressed, e.g. for a compressed
ramdisk/initrd which guests can unpack themselves.
For this reason, we only uncompress when requested
via the new 'uncompress' option.
Uncompressing is still useful, for example, to save
space on the local disk.

Requires-Change-Id: I2fb0c5c99cb899cf50f8753aa2720e2247a1ef7b
Requires-Change-Id: I80565199427e7266f3f47ab2c956c6beaec12f0b

+++ server/src/build.pl
@@ -100,27 +100,29 @@ my %output_formatter = (
 # build object files from the modules
 sub build_obj
 {
-  my ($file, $cmdline, $modname, $flags, $no_strip) = @_;
+  my ($file, $cmdline, $modname, $flags, $opts) = @_;
   my %d;
   my %imgmod;
 
   $d{path} = L4::ModList::search_file($file, $module_path)
     || die "Cannot find file $file! Used search path: $module_path";
 
-  # make sure that the file isn't already compressed
-  system("$prog_gzip -dc $d{path} > $modname.ugz 2> /dev/null");
-  if ($?)
+  if (exists $opts->{uncompress})
     {
-      unlink("$modname.ugz");
-    }
-  else
-    {
-      $d{path} = "$modname.ugz";
+      system("$prog_gzip -dc $d{path} > $modname.ugz 2> /dev/null");
+      if ($?)
+        {
+          unlink("$modname.ugz");
+        }
+      else
+        {
+          $d{path} = "$modname.ugz";
+        }
     }
   $d{size_orig} = -s $d{path};
 
   my $take_orig = 1;
-  if ($strip and not $no_strip)
+  if ($strip and not exists $opts->{nostrip})
     {
       system("$prog_objcopy -S $d{path} $modname.obj 2> /dev/null");
       $take_orig = $?;
@@ -140,10 +142,11 @@ sub build_obj
   $c_unc->addfile(*M);
   close M;
 
-  if ($compress)
+  if ($compress and
+      not L4::ModList::is_gzipped_file("$modname.obj"))
     {
-      system("$prog_gzip -9f $modname.obj && mv $modname.obj.gz $modname.obj");
-      $d{size_compressed} = -s "$modname.obj";
+       system("$prog_gzip -9f $modname.obj && mv $modname.obj.gz $modname.obj");
+       $d{size_compressed} = -s "$modname.obj";
     }
 
   $d{modname} = $modname;
@@ -209,7 +212,7 @@ sub build_objects(@)
     $img{mods}[$i] =
       { build_obj($mods[$i]->{command}, $mods[$i]->{cmdline},
                   $mods[$i]->{modname}, $flags,
-                  $mods[$i]->{type} =~ /.+-nostrip$/) };
+                  $mods[$i]->{opts}) };
   }
 
   &{$output_formatter{end}}();
